Experiment using esbuild for building, which should be faster. (Although generating the type definitions file - index.d.ts - is still done with tsc so maybe it doesn't really help).
In order to make esbuild work, we added esModuleInterop: true to tsconig.json.
As a result of this, many 'import * as thing' had to be rewritten as
'import thing' to make things work again.

We made chages to package.json, added a build:esbuild step that runs
'node esbuild.config.js'
esbuild.config.js file also takes care of making the dist folder empty first
and copying js/docs to the dist folder.
esbuild also takes care of bundling and minifying the code into both an
ESM and CJS module.
We added an 'exports' section to package.json to point to both versions.

Now we need to test if this can be installed or used inside a CJS as well as an ESM
project.